unify normal/param/assume tests
