--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "Ext/libhandler"]
-	path = Ext/libhandler
-	url = https://github.com/koka-lang/libhandler.git
